refactor: move document_to_element_list from common module to pdf module
